2026-04-17-chainguard-ai/requirements.txt
streamlit>=1.44.0
pandas>=2.2.3
plotly>=5.24.1
pydeck>=0.9.1
requests>=2.32.3
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic>=2.11.0
numpy>=2.2.4
scikit-learn>=1.6.1
langchain>=0.3.24
langchain-core>=0.3.56
langgraph>=0.3.34
langchain-groq>=0.3.2
hindsight-client>=0.4.0
hindsight-langgraph[langgraph]>=0.1.1
gspread>=6.1.4
google-auth>=2.39.0
google-api-python-client>=2.168.0
tenacity>=9.1.2
pyttsx3>=2.98
